requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
crispy-bootstrap4==2024.1
cryptography==42.0.8
cssselect2==0.7.0
dacktool==0.0.7
dbstream==0.1.25
defusedxml==0.8.0rc2
dj-database-url==2.2.0
Django==5.0.6
django-allauth==0.63.3
django-cors-headers==4.4.0
django-countries==7.6.1
django-crispy-forms==2.3
django-dotenv==1.4.2
django-nested-admin==4.1.1
django-redis==5.4.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
easyocr==1.7.1
et_xmlfile==2.0.0
filelock==3.16.0
fonttools==4.53.1
fpdf2==2.7.9
fsspec==2024.9.0
google-api-core==2.19.1
google-api-python-client==1.7.11
google-auth==2.31.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==1.2.0
google-cloud-secret-manager==2.7.2
googleapis-common-protos==1.63.2
googleauthentication==0.0.17
grpc-google-iam-v1==0.12.7
grpcio==1.64.1
grpcio-status==1.62.2
gunicorn==23.0.0
html5lib==1.1
httplib2==0.22.0
idna==3.7
imageio==2.35.1
Jinja2==3.1.4
lazy_loader==0.4
libcst==1.4.0
lxml==5.2.2
MarkupSafe==2.1.5
mpmath==1.3.0
mysqlclient==2.2.4
networkx==3.3
ninja==1.11.1.1
numpy==2.1.1
oauthlib==3.2.2
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
oscrypto==1.3.0
packaging==24.1
pandas==2.3.1
pillow==10.3.0
proto-plus==1.24.0
protobuf==4.25.3
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyclipper==1.3.0.post5
pycparser==2.22
pydyf==0.11.0
pyHanko==0.25.0
pyhanko-certvalidator==0.26.3
PyJWT==2.10.1
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pyparsing==3.1.2
pypdf==4.2.0
pyphen==0.16.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-http-client==3.3.7
python-monkey-business==1.1.0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.1
qrcode==7.4.2
razorpay==1.4.2
redis==5.2.1
reportlab==4.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-image==0.24.0
scipy==1.14.1
setuptools==74.1.2
shapely==2.0.6
six==1.16.0
social-auth-core==4.5.4
sqlparse==0.5.0
starkbank-ecdsa==2.2.0
svglib==1.5.1
sympy==1.13.2
tifffile==2024.8.30
tinycss2==1.3.0
torch==2.4.1
torchvision==0.19.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==3.0.1
uritools==4.0.3
urllib3==2.2.2
weasyprint==62.3
webencodings==0.5.1
xhtml2pdf==0.2.16
zopfli==0.2.3
